evil behavior, moving the cursor when reaching the end of the line
https://github.com/emacs-evil/evil/commit/aa3ea1dcfc54fe25581841a8685679d482e99b3f

Unwed evil-move-beyond-eol and evil-move-cursor-back

The docstring of evil-cursor-back only mentions that it applies to moving the
cursor back when exiting insert state, but in many cases it was tied to also
moving the cursor when reaching the end of the line. We have
evil-move-beyond-eol to control that behavior, and there doesn't seem to be a
good reason that the two must be tied together.